initial_stress refactor
This is in preparation for deprecation of the initial_stress parameter in MOOSE's ComputeStressBase class.  In the future, MOOSE will default to using ComputeEigenstrainFromInitialStress instead of specifying the initial_stress parameter in ComputeStressBase, as this is more flexible when using time-varying elasticity tensors.  You can see an example of how to use this in ex04/2BlockFriction_Isoilunit.i

However, the ComputeISoilStress uses stress_old in its algorithm, and I'm not confident that I can successfully change the algorithm to use the old elastic strain, as we've done for MOOSE's TensorMechanics.  Therefore, I've created a new input parameter called "initial_soil_stress" that exactly replaces ComputeStressBase's "initial_stress".  Then ComputeISoilStress can happily use stress_old, and ComputeStressBase can deprecate its initial_stress.

The remainder of the changes are just the relevant changes in the Action and the test files.

Fixes #85