removed boost distribution calls.  Calling regular lognormal distribution static functions ref #309
- In calcLogLikelihood  set lognormal::cdf to numeric_limits::min when it returns 0.
Boost would return a small number.  This was making a unit test fail

I didn't create any distribution objects so I changed the interface to greaterProbability to take
the parameters needed to create the lognormal distribution.  If we want to create these distribution
objects, we will need to change the input file or create them in the fragility constructor.
greaterProbability is fixed to only deal with lognormal distributions and not general distributions which
would require an distribution object to be created.